qcreport: relocate the docwriter.py module up a level.
